pgsql

+++ database/migrations/2025_06_29_125757_add_sensor_name_to_sensor_lists_table.php
@@ -11,15 +11,22 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('sensor_lists', function (Blueprint $table) {
-            $table->string('sensor_name', 255)->after('blink_status');
-        });
+        if (! Schema::hasColumn('sensor_lists', 'sensor_name')) {
+            Schema::table('sensor_lists', function (Blueprint $table) {
+                $table->string('sensor_name', 255)->after('blink_status');
+            });
+        }
     }
 
+    /**
+     * Reverse the migrations.
+     */
     public function down(): void
     {
-        Schema::table('sensor_lists', function (Blueprint $table) {
-            $table->dropColumn('sensor_name');
-        });
+        if (Schema::hasColumn('sensor_lists', 'sensor_name')) {
+            Schema::table('sensor_lists', function (Blueprint $table) {
+                $table->dropColumn('sensor_name');
+            });
+        }
     }
 };
